// lib/generated/serviceEnums.ts
// 🚨 AUTO-GENERATED FILE. DO NOT EDIT MANUALLY!
// Generated from Prisma schema.


export enum ServiceCategory {
  MANUAL_THERAPY = "MANUAL_THERAPY",
  CONSULTATION = "CONSULTATION",
  ELECTROTHERAPY = "ELECTROTHERAPY",
  EXERCISE_THERAPY = "EXERCISE_THERAPY",
  COMBO_TREATMENT = "COMBO_TREATMENT",
}